pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "nonebot-plugin-simple-setu"
version = "0.2.0"
description = "一个简单到不能再简单的色图插件"
readme = "README.md"
requires-python = ">=3.10,<4.0"
license = { file = "LICENSE" }

authors = [
    { name = "nomdn", email = "iduhih777@outlook.com" }
]

keywords = ["nonebot2", "plugin", "setu", "image", "pixiv"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: GNU Lesser General Public License v3 (LGPLv3)",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Utilities",
]

dependencies = [
    "nonebot2 >= 2.3.0,<3.0.0",
    "nonebot-adapter-onebot >= 2.0.0",
    "httpx >= 0.28.1",
    "nonebot-plugin-send-anything-anywhere>=0.7.1",
    "nonebot-plugin-limiter>=0.4.3",
    "tomli-w>=1.2.0",
]

[project.urls]
homepage = "https://github.com/nomdn/nonebot-plugin-simple-setu"
repository = "https://github.com/nomdn/nonebot-plugin-simple-setu"
documentation = "https://github.com/nomdn/nonebot-plugin-simple-setu"

[project.optional-dependencies]
dev = [
    "pytest",
    "black",
    "flake8",
    "isort",
]
[[tool.uv.index]]
name = "tencent"
url = "https://pypi.tuna.tsinghua.edu.cn/simple/"
